backend/requirements.txt
fastapi
uvicorn
python-dotenv

langchain
langchain-core
langchain-community
langchain-google-genai

pydantic
pypdf

python-docx
pillow

python-multipart